move generateAndAddTsePdf to tseService
